pyproject.toml
[project]
name = "connectmap"
version = "0.1.0"
description = "A simple web application to visualize Garmin Connect activities on OpenStreetMap"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "garminconnect>=0.2.13",
    "python-dotenv>=1.0.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "codespell",
    "pre-commit",
    "prettier",
    "ruff",
]

[tool.uv]
required-version = ">=0.9.13"

[tool.ruff]
line-length = 160
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.codespell]
skip = '*.http,*.json'
